# swagger/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Adventure APIs",
    "description": "API for admin/managing and client/viewing",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/friend/browse": {
      "get": {
        "tags": [
          "friend"
        ],
        "description": "list friend of an user",
        "operationId": "list_friend",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page to pagination",
            "required": false,
            "type": "number"
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "item per page for pagination",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4 : Friend is not found; ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "first_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "last_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "password": {
                      "type": "string",
                      "example": "123456"
                    },
                    "email": {
                      "type": "string",
                      "example": "Brucelee Thanh@gmail.com"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": ""
                    },
                    "gender": {
                      "type": "number",
                      "example": 0
                    },
                    "birthday": {
                      "type": "date-time",
                      "example": "1993-07-05T04:27:58.144Z"
                    },
                    "address": {
                      "type": "string",
                      "example": ""
                    },
                    "religion": {
                      "type": "string",
                      "example": ""
                    },
                    "intro": {
                      "type": "string",
                      "example": ""
                    },
                    "fb_id": {
                      "type": "string",
                      "example": ""
                    },
                    "avatar": {
                      "type": "string",
                      "example": ""
                    },
                    "cover": {
                      "type": "string",
                      "example": ""
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "last_visited_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    }
                  }
                },
                "total": {
                  "type": "number",
                  "example": 1
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/friend/search": {
      "post": {
        "tags": [
          "friend"
        ],
        "description": "search friend or other people by keyword",
        "operationId": "search_friend",
        "parameters": [
          {
            "name": "token",
            "in": "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "keyword",
            "in": "formData",
            "description": "keyword",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "formData",
            "description": "page to pagination",
            "required": false,
            "type": "number"
          },
          {
            "name": "per_page",
            "in": "formData",
            "description": "item per page for pagination",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4 : Can not found anythings; ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "friends":{
                      "type": "array",
                      "item":{
                        "_id": {
                          "type": "string",
                          "example": "5769097529f87b5c271728e5"
                        },
                        "first_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "last_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "avatar": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "total_friend:": {
                      "type": "number",
                      "example": 10
                    },
                    "strangers":{
                      "type": "array",
                      "item":{
                        "_id": {
                          "type": "string",
                          "example": "5769097529f87b5c271728e5"
                        },
                        "first_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "last_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "avatar": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "total_stranger:": {
                      "type": "number",
                      "example": 10
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/friend/suggest_friend": {
      "get": {
        "tags": [
          "friend"
        ],
        "description": "suggest friend",
        "operationId": "suggest_friend",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page to pagination",
            "required": false,
            "type": "number"
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "item per page for pagination",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4 : User is not found;   ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "first_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "last_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "password": {
                      "type": "string",
                      "example": "123456"
                    },
                    "email": {
                      "type": "string",
                      "example": "Brucelee Thanh@gmail.com"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": ""
                    },
                    "gender": {
                      "type": "number",
                      "example": 0
                    },
                    "birthday": {
                      "type": "date-time",
                      "example": "1993-07-05T04:27:58.144Z"
                    },
                    "address": {
                      "type": "string",
                      "example": ""
                    },
                    "religion": {
                      "type": "string",
                      "example": ""
                    },
                    "intro": {
                      "type": "string",
                      "example": ""
                    },
                    "fb_id": {
                      "type": "string",
                      "example": ""
                    },
                    "avatar": {
                      "type": "string",
                      "example": ""
                    },
                    "cover": {
                      "type": "string",
                      "example": ""
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "last_visited_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    }
                  }
                },
                "total": {
                  "type": "number",
                  "example": 1
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}